+++ core/src/main/scala/org/http4s/multipart/MultipartParser.scala
@@ -5,25 +5,36 @@ import cats.effect._
 import cats.implicits.{catsSyntaxEither => _, _}
 import fs2._
 
+import java.nio.file._
+
+import java.nio.file._
+
 /** A low-level multipart-parsing pipe.  Most end users will prefer EntityDecoder[Multipart]. */
 object MultipartParser {
 
-  private val CRLFBytesN = Array[Byte]('\r', '\n')
-  private val DoubleCRLFBytesN = Array[Byte]('\r', '\n', '\r', '\n')
-  private val DashDashBytesN = Array[Byte]('-', '-')
-  private val BoundaryBytesN: Boundary => Array[Byte] = boundary => boundary.value.getBytes("UTF-8")
+  private[this] val logger = org.log4s.getLogger
+
+  private[this] val CRLFBytesN = Array[Byte]('\r', '\n')
+  private[this] val DoubleCRLFBytesN = Array[Byte]('\r', '\n', '\r', '\n')
+  private[this] val DashDashBytesN = Array[Byte]('-', '-')
+  private[this] val BoundaryBytesN: Boundary => Array[Byte] = boundary =>
+    boundary.value.getBytes("UTF-8")
   val StartLineBytesN: Boundary => Array[Byte] = BoundaryBytesN.andThen(DashDashBytesN ++ _)
 
-  private val ExpectedBytesN: Boundary => Array[Byte] =
+  private[this] val ExpectedBytesN: Boundary => Array[Byte] =
     BoundaryBytesN.andThen(CRLFBytesN ++ DashDashBytesN ++ _)
+  private[this] val dashByte: Byte = '-'.toByte
+  private[this] val streamEmpty = Stream.empty
+  private[this] val PullUnit = Pull.pure[Pure, Unit](())
 
-  private val EndlineBytesN: Boundary => Array[Byte] =
-    BoundaryBytesN.andThen(CRLFBytesN ++ DashDashBytesN ++ _ ++ DashDashBytesN)
+  private type SplitStream[F[_]] = Pull[F, Nothing, (Stream[F, Byte], Stream[F, Byte])]
+  private type SplitFileStream[F[_]] =
+    Pull[F, Nothing, (Stream[F, Byte], Stream[F, Byte], Option[Path])]
 
   def parseStreamed[F[_]: Sync](
       boundary: Boundary,
       limit: Int = 1024): Pipe[F, Byte, Multipart[F]] = { st =>
-    ignorePreludeStage[F](boundary, st, limit)
+    ignorePrelude[F](boundary, st, limit)
       .fold(Vector.empty[Part[F]])(_ :+ _)
       .map(Multipart(_, boundary))
   }
@@ -31,7 +42,7 @@ object MultipartParser {
   def parseToPartsStream[F[_]: Sync](
       boundary: Boundary,
       limit: Int = 1024): Pipe[F, Byte, Part[F]] = { st =>
-    ignorePreludeStage[F](boundary, st, limit)
+    ignorePrelude[F](boundary, st, limit)
   }
 
   private def splitAndIgnorePrev[F[_]](
@@ -97,17 +108,9 @@ object MultipartParser {
 
   /** Split a chunk in the case of a partial match:
     *
-    * If it is a chunk that is between a partial match
-    * (middle chunked), the prior partial match is added to
-    * the accumulator, and the current partial match is
-    * considered to carry over.
-    *
-    * If it is a fresh chunk (no carry over partial match),
-    * everything prior to the partial match is added to the accumulator,
-    * and the partial match is considered the carry over.
-    *
-    * Else, if the whole block is a partial match,
-    * add it to the carry over
+    * DO NOT USE. Was made private[http4s] because
+    * Jose messed up hard like 5 patches ago and now it breaks bincompat to
+    * remove.
     *
     */
   private def splitPartialMatch[F[_]: Sync](
@@ -141,7 +144,7 @@ object MultipartParser {
     * from the subsequent split stream).
     *
     */
-  private def splitOnChunk[F[_]: Sync](
+  private[http4s] def splitOnChunk[F[_]: Sync](
       values: Array[Byte],
       state: Int,
       c: Chunk[Byte],
@@ -177,10 +180,10 @@ object MultipartParser {
 
   /** The first part of our streaming stages:
     *
-    * Ignore the prelude and remove the first boundary
-    *
+    * Ignore the prelude and remove the first boundary. Only traverses until the first
+    * part
     */
-  private def ignorePreludeStage[F[_]: Sync](
+  private[this] def ignorePrelude[F[_]: Sync](
       b: Boundary,
       stream: Stream[F, Byte],
       limit: Int): Stream[F, Part[F]] = {
@@ -188,12 +191,11 @@ object MultipartParser {
 
     def go(s: Stream[F, Byte], state: Int, strim: Stream[F, Byte]): Pull[F, Part[F], Unit] =
       if (state == values.length) {
-        streamStageIgnoreRest(b, strim ++ s, limit).pull.echo
+        pullParts[F](b, strim ++ s, limit)
       } else {
         s.pull.unconsChunk.flatMap {
           case Some((chnk, rest)) =>
-            val bytes = chnk
-            val (ix, strim) = splitAndIgnorePrev(values, state, bytes)
+            val (ix, strim) = splitAndIgnorePrev(values, state, chnk)
             go(rest, ix, strim)
           case None =>
             Pull.raiseError(MalformedMessageBodyFailure("Malformed Malformed match"))
@@ -209,141 +211,127 @@ object MultipartParser {
     }.stream
   }
 
-  private def parseToPartStreamed[F[_]: Sync](s: Stream[F, Byte], limit: Int): Stream[F, Part[F]] =
-    splitLimited[F](DoubleCRLFBytesN, s, limit).flatMap {
-      case (l, r) =>
-        l.pull.uncons.flatMap {
-          case None =>
-            Pull.raiseError(
-              MalformedMessageBodyFailure("Invalid Separation between headers and body"))
-          case Some(_) =>
-            parseHeaders[F](l).map(Part[F](_, r)).pull.echo *> Pull.done
-        }
-    }.stream
-
-  private def parseHeaders[F[_]: Sync](strim: Stream[F, Byte]): Stream[F, Headers] = {
-    def tailrecParse(s: Stream[F, Byte], headers: Headers): Pull[F, Headers, Unit] =
-      splitHalf[F](CRLFBytesN, s).flatMap {
-        case (l, r) =>
-          l.through(fs2.text.utf8Decode[F])
-            .fold("")(_ ++ _)
-            .map { string =>
-              val ix = string.indexOf(':')
-              if (string.indexOf(':') >= 0)
-                headers.put(Header(string.substring(0, ix), string.substring(ix + 1).trim))
-              else
-                headers
-            }
-            .pull
-            .echo *> r.pull.uncons.flatMap {
-            case Some(_) =>
-              tailrecParse(r, headers)
-            case None =>
-              Pull.done
-          }
-      }
-
-    tailrecParse(strim, Headers.empty).stream
-      .fold(Headers.empty)(_ ++ _)
-  }
-
-  private def streamStageIgnoreRest[F[_]: Sync](
+  /**
+    *
+    * @param boundary
+    * @param s
+    * @param limit
+    * @tparam F
+    * @return
+    */
+  private def pullParts[F[_]: Sync](
       boundary: Boundary,
       s: Stream[F, Byte],
       limit: Int
-  ): Stream[F, Part[F]] = {
-    val endlineBytes = EndlineBytesN(boundary)
-    val values = ExpectedBytesN(boundary)
-    splitOrFail[F](endlineBytes, s).flatMap {
-      case (l, _) =>
-        streamStageParsePart[F](boundary, values, l ++ Stream.chunk(Chunk.bytes(values)), limit).pull.echo
-    }.stream
-  }
+  ): Pull[F, Part[F], Unit] = {
+    val values = DoubleCRLFBytesN
+    val expectedBytes = ExpectedBytesN(boundary)
 
-  private def streamStageParsePart[F[_]: Sync](
-      boundary: Boundary,
-      values: Array[Byte],
-      s: Stream[F, Byte],
-      limit: Int
-  ): Stream[F, Part[F]] =
-    splitHalf[F](values, s).flatMap {
+    splitOrFinish[F](values, s, limit).flatMap {
       case (l, r) =>
-        r.pull.unconsChunk.flatMap {
-          case None =>
-            parseToPartStreamed[F](l, limit).pull.echo *>
-              Pull.done
-          case Some(_) =>
-            tailrecParts[F](boundary, values, l, r, limit)
+        //We can abuse reference equality here for efficiency
+        //Since `splitOrFinish` returns `empty` on a capped stream
+        //However, we must have at least one part, so `splitOrFinish` on this function
+        //Indicates an error
+        if (r == streamEmpty) {
+          Pull.raiseError(MalformedMessageBodyFailure("Cannot parse empty stream"))
+        } else {
+          tailrecParts[F](boundary, l, r, expectedBytes, limit)
         }
-    }.stream
+    }
+  }
 
   private def tailrecParts[F[_]: Sync](
       b: Boundary,
-      values: Array[Byte],
-      last: Stream[F, Byte],
-      next: Stream[F, Byte],
-      limit: Int
-  ): Pull[F, Part[F], Unit] =
-    parseToPartStreamed[F](last, limit).pull.echo *> splitHalf[F](values, next).flatMap {
-      case (l, r) =>
-        r.pull.uncons.flatMap {
-          case None =>
-            Pull.done
-          case Some(_) =>
-            tailrecParts[F](b, values, l, r, limit)
-        }
-    }
-
-  private def splitHalf[F[_]: Sync](
-      values: Array[Byte],
-      stream: Stream[F, Byte]): Pull[F, Nothing, (Stream[F, Byte], Stream[F, Byte])] = {
-
-    def go(
-        s: Stream[F, Byte],
-        state: Int,
-        lacc: Stream[F, Byte],
-        racc: Stream[F, Byte]): Pull[F, Nothing, (Stream[F, Byte], Stream[F, Byte])] =
-      if (state == values.length) {
-        Pull.pure((lacc, racc ++ s))
-      } else {
-        s.pull.unconsChunk.flatMap {
-          case Some((chnk, str)) =>
-            val (ix, l, r) = splitOnChunk[F](values, state, chnk, lacc, racc)
-            go(str, ix, l, r)
-          case None =>
-            if (state != 0) {
-              Pull.raiseError(MalformedMessageBodyFailure("Invalid boundary - partial boundary"))
+      headerStream: Stream[F, Byte],
+      rest: Stream[F, Byte],
+      expectedBytes: Array[Byte],
+      limit: Int): Pull[F, Part[F], Unit] =
+    Pull
+      .eval(parseHeaders(headerStream))
+      .flatMap { hdrs =>
+        splitHalf(expectedBytes, rest).flatMap {
+          case (l, r) =>
+            //We hit a boundary, but the rest of the stream is empty
+            //and thus it's not a properly capped multipart body
+            if (r == streamEmpty) {
+              Pull.raiseError(MalformedMessageBodyFailure("Part not terminated properly"))
             } else {
-              Pull.pure((lacc, racc))
+              Pull.output1(Part[F](hdrs, l)) >> splitOrFinish(DoubleCRLFBytesN, r, limit).flatMap {
+                case (hdrStream, remaining) =>
+                  if (hdrStream == streamEmpty) { //Empty returned if it worked fine
+                    Pull.done
+                  } else {
+                    tailrecParts[F](b, hdrStream, remaining, expectedBytes, limit)
+                  }
+              }
             }
         }
       }
 
-    stream.pull.unconsChunk.flatMap {
-      case Some((chunk, rest)) =>
-        val (ix, l, r) = splitOnChunk[F](values, 0, chunk, Stream.empty, Stream.empty)
-        go(rest, ix, l, r)
-      case None =>
-        Pull.pure((Stream.empty, Stream.empty))
-    }
-  }
-
-  private def splitOrFail[F[_]: Sync](
+  /** Split a stream in half based on `values`,
+    * but check if it is either double dash terminated (end of multipart).
+    * SplitOrFinish also tracks a header limit size
+    *
+    * If it is, return the empty stream. if it is not, split on the `values`
+    * and raise an error if we lack a match
+    */
+  //noinspection ScalaStyle
+  private def splitOrFinish[F[_]: Sync](
       values: Array[Byte],
-      stream: Stream[F, Byte]): Pull[F, Nothing, (Stream[F, Byte], Stream[F, Byte])] = {
+      stream: Stream[F, Byte],
+      limit: Int): SplitStream[F] = {
+
+    //Check if a particular chunk a final chunk, that is,
+    //whether it's the boundary plus an extra "--", indicating it's
+    //the last boundary
+    def checkIfLast(c: Chunk[Byte], rest: Stream[F, Byte]): SplitStream[F] =
+      if (c.size <= 0) {
+        Pull.raiseError(MalformedMessageBodyFailure("Invalid Chunk: Chunk is empty"))
+      } else if (c.size == 1) {
+        rest.pull.unconsChunk.flatMap {
+          case Some((chnk, remaining)) =>
+            if (chnk.size <= 0)
+              Pull.raiseError(MalformedMessageBodyFailure("Invalid Chunk: Chunk is empty"))
+            else if (c(0) == dashByte && chnk(0) == dashByte) {
+              Pull.pure((streamEmpty, streamEmpty))
+            } else {
+              val (ix, l, r, add) =
+                splitOnChunkLimited[F](
+                  values,
+                  0,
+                  Chunk.bytes(c.toArray[Byte] ++ chnk.toArray[Byte]),
+                  Stream.empty,
+                  Stream.empty)
+              go(remaining, ix, l, r, add)
+            }
+          case None =>
+            Pull.raiseError(MalformedMessageBodyFailure("Malformed Multipart ending"))
+        }
+      } else if (c(0) == dashByte && c(1) == dashByte) {
+        Pull.pure((streamEmpty, streamEmpty))
+      } else {
+        val (ix, l, r, add) =
+          splitOnChunkLimited[F](values, 0, c, Stream.empty, Stream.empty)
+        go(rest, ix, l, r, add)
+      }
 
     def go(
         s: Stream[F, Byte],
         state: Int,
         lacc: Stream[F, Byte],
-        racc: Stream[F, Byte]): Pull[F, Nothing, (Stream[F, Byte], Stream[F, Byte])] =
-      if (state == values.length) {
+        racc: Stream[F, Byte],
+        limitCTR: Int): SplitStream[F] =
+      if (limitCTR >= limit) {
+        Pull.raiseError(
+          MalformedMessageBodyFailure(s"Part header was longer than $limit-byte limit"))
+      } else if (state == values.length) {
         Pull.pure((lacc, racc ++ s))
       } else {
         s.pull.unconsChunk.flatMap {
           case Some((chnk, str)) =>
-            val (ix, l, r) = splitOnChunk[F](values, state, chnk, lacc, racc)
-            go(str, ix, l, r)
+            val (ix, l, r, add) = splitOnChunkLimited[F](values, state, chnk, lacc, racc)
+            go(str, ix, l, r, limitCTR + add)
           case None =>
             Pull.raiseError(MalformedMessageBodyFailure("Invalid boundary - partial boundary"))
         }
@@ -351,10 +339,76 @@ object MultipartParser {
 
     stream.pull.unconsChunk.flatMap {
       case Some((chunk, rest)) =>
+        checkIfLast(chunk, rest)
+      case None =>
+        Pull.raiseError(MalformedMessageBodyFailure("Invalid boundary - partial boundary"))
+    }
+  }
+
+  /** Take the stream of headers separated by
+    * double CRLF bytes and return the headers
+    */
+  private def parseHeaders[F[_]: Sync](strim: Stream[F, Byte]): F[Headers] = {
+    def tailrecParse(s: Stream[F, Byte], headers: Headers): Pull[F, Headers, Unit] =
+      splitHalf[F](CRLFBytesN, s).flatMap {
+        case (l, r) =>
+          l.through(fs2.text.utf8Decode[F])
+            .fold("")(_ ++ _)
+            .map { string =>
+              val ix = string.indexOf(':')
+              if (ix >= 0) {
+                headers.put(Header(string.substring(0, ix), string.substring(ix + 1).trim))
+              } else {
+                headers
+              }
+            }
+            .pull
+            .echo >> r.pull.uncons.flatMap {
+            case Some(_) =>
+              tailrecParse(r, headers)
+            case None =>
+              Pull.done
+          }
+      }
+
+    tailrecParse(strim, Headers.empty).stream.compile
+      .fold(Headers.empty)(_ ++ _)
+  }
+
+  /** Spit our `Stream[F, Byte]` into two halves.
+    * If we reach the end and the state is 0 (meaning we didn't match at all),
+    * then we return the concatenated parts of the stream.
+    *
+    * This method _always_ caps
+    */
+  private def splitHalf[F[_]: Sync](
+      values: Array[Byte],
+      stream: Stream[F, Byte]): SplitStream[F] = {
+
+    def go(
+        s: Stream[F, Byte],
+        state: Int,
+        lacc: Stream[F, Byte],
+        racc: Stream[F, Byte]): SplitStream[F] =
+      if (state == values.length) {
+        Pull.pure((lacc, racc ++ s))
+      } else {
+        s.pull.unconsChunk.flatMap {
+          case Some((chnk, str)) =>
+            val (ix, l, r) = splitOnChunk[F](values, state, chnk, lacc, racc)
+            go(str, ix, l, r)
+          case None =>
+            //We got to the end, and matched on nothing.
+            Pull.pure((lacc ++ racc, streamEmpty))
+        }
+      }
+
+    stream.pull.unconsChunk.flatMap {
+      case Some((chunk, rest)) =>
         val (ix, l, r) = splitOnChunk[F](values, 0, chunk, Stream.empty, Stream.empty)
         go(rest, ix, l, r)
       case None =>
-        Pull.raiseError(MalformedMessageBodyFailure("Invalid boundary - partial boundary"))
+        Pull.pure((streamEmpty, streamEmpty))
     }
   }
 
@@ -411,7 +465,7 @@ object MultipartParser {
     * add it to the carry over
     *
     */
-  private def splitPartialLimited[F[_]: Sync](
+  private[http4s] def splitPartialLimited[F[_]: Sync](
       state: Int,
       middleChunked: Boolean,
       currState: Int,
@@ -434,7 +488,7 @@ object MultipartParser {
     }
   }
 
-  private def splitOnChunkLimited[F[_]: Sync](
+  private[http4s] def splitOnChunkLimited[F[_]: Sync](
       values: Array[Byte],
       state: Int,
       c: Chunk[Byte],
@@ -469,33 +523,253 @@ object MultipartParser {
     }
   }
 
-  private def splitLimited[F[_]: Sync](
+  ////////////////////////////////////////////////////////////
+  // File writing encoder
+  ///////////////////////////////////////////////////////////
+
+  /** Same as the other streamed parsing, except
+    * after a particular size, it buffers on a File.
+    */
+  def parseStreamedFile[F[_]: Sync](
+      boundary: Boundary,
+      limit: Int = 1024,
+      maxSizeBeforeWrite: Int = 52428800,
+      maxParts: Int = 20,
+      failOnLimit: Boolean = false): Pipe[F, Byte, Multipart[F]] = { st =>
+    ignorePreludeFileStream[F](boundary, st, limit, maxSizeBeforeWrite, maxParts, failOnLimit)
+      .fold(Vector.empty[Part[F]])(_ :+ _)
+      .map(Multipart(_, boundary))
+  }
+
+  def parseToPartsStreamedFile[F[_]: Sync](
+      boundary: Boundary,
+      limit: Int = 1024,
+      maxSizeBeforeWrite: Int = 52428800,
+      maxParts: Int = 20,
+      failOnLimit: Boolean = false): Pipe[F, Byte, Part[F]] = { st =>
+    ignorePreludeFileStream[F](boundary, st, limit, maxSizeBeforeWrite, maxParts, failOnLimit)
+  }
+
+  /** The first part of our streaming stages:
+    *
+    * Ignore the prelude and remove the first boundary. Only traverses until the first
+    * part
+    */
+  private[this] def ignorePreludeFileStream[F[_]: Sync](
+      b: Boundary,
+      stream: Stream[F, Byte],
+      limit: Int,
+      maxSizeBeforeWrite: Int,
+      maxParts: Int,
+      failOnLimit: Boolean): Stream[F, Part[F]] = {
+    val values = StartLineBytesN(b)
+
+    def go(s: Stream[F, Byte], state: Int, strim: Stream[F, Byte]): Pull[F, Part[F], Unit] =
+      if (state == values.length) {
+        pullPartsFileStream[F](b, strim ++ s, limit, maxSizeBeforeWrite, maxParts, failOnLimit)
+      } else {
+        s.pull.unconsChunk.flatMap {
+          case Some((chnk, rest)) =>
+            val (ix, strim) = splitAndIgnorePrev(values, state, chnk)
+            go(rest, ix, strim)
+          case None =>
+            Pull.raiseError(MalformedMessageBodyFailure("Malformed Malformed match"))
+        }
+      }
+
+    stream.pull.unconsChunk.flatMap {
+      case Some((chnk, strim)) =>
+        val (ix, rest) = splitAndIgnorePrev(values, 0, chnk)
+        go(strim, ix, rest)
+      case None =>
+        Pull.raiseError(MalformedMessageBodyFailure("Cannot parse empty stream"))
+    }.stream
+  }
+
+  /**
+    *
+    * @param boundary
+    * @param s
+    * @param limit
+    * @tparam F
+    * @return
+    */
+  private def pullPartsFileStream[F[_]: Sync](
+      boundary: Boundary,
+      s: Stream[F, Byte],
+      limit: Int,
+      maxBeforeWrite: Int,
+      maxParts: Int,
+      failOnLimit: Boolean
+  ): Pull[F, Part[F], Unit] = {
+    val values = DoubleCRLFBytesN
+    val expectedBytes = ExpectedBytesN(boundary)
+
+    splitOrFinish[F](values, s, limit).flatMap {
+      case (l, r) =>
+        //We can abuse reference equality here for efficiency
+        //Since `splitOrFinish` returns `empty` on a capped stream
+        //However, we must have at least one part, so `splitOrFinish` on this function
+        //Indicates an error
+        if (r == streamEmpty) {
+          Pull.raiseError(MalformedMessageBodyFailure("Cannot parse empty stream"))
+        } else {
+          tailrecPartsFileStream[F](
+            boundary,
+            l,
+            r,
+            expectedBytes,
+            limit,
+            maxBeforeWrite,
+            1,
+            maxParts,
+            failOnLimit
+          )
+        }
+    }
+  }
+
+  private[this] def cleanupFileOption[F[_]](p: Option[Path])(
+      implicit F: Sync[F]): Pull[F, Nothing, Unit] =
+    p match {
+      case Some(path) =>
+        Pull.eval(cleanupFile(path))
+
+      case None =>
+        PullUnit //Todo: Move to fs2
+    }
+
+  private[this] def cleanupFile[F[_]](path: Path)(implicit F: Sync[F]): F[Unit] =
+    F.delay(Files.delete(path))
+      .handleErrorWith { err =>
+        logger.error(err)("Caught error during file cleanup for multipart")
+        //Swallow and report io exceptions in case
+        F.unit
+      }
+
+  private[this] def tailrecPartsFileStream[F[_]: Sync](
+      b: Boundary,
+      headerStream: Stream[F, Byte],
+      rest: Stream[F, Byte],
+      expectedBytes: Array[Byte],
+      headerLimit: Int,
+      maxBeforeWrite: Int,
+      partsCounter: Int,
+      partsLimit: Int,
+      failOnLimit: Boolean): Pull[F, Part[F], Unit] =
+    Pull
+      .eval(parseHeaders(headerStream))
+      .flatMap { hdrs =>
+        splitWithFileStream(expectedBytes, rest, maxBeforeWrite).flatMap {
+          case (partBody, rest, fileRef) =>
+            //We hit a boundary, but the rest of the stream is empty
+            //and thus it's not a properly capped multipart body
+            if (rest == streamEmpty) {
+              cleanupFileOption[F](fileRef) >> Pull.raiseError(
+                MalformedMessageBodyFailure("Part not terminated properly"))
+            } else {
+              Pull.output1(makePart(hdrs, partBody, fileRef)) >> splitOrFinish(
+                DoubleCRLFBytesN,
+                rest,
+                headerLimit)
+                .flatMap {
+                  case (hdrStream, remaining) =>
+                    if (hdrStream == streamEmpty) { //Empty returned if it worked fine
+                      Pull.done
+                    } else if (partsCounter >= partsLimit) {
+                      if (failOnLimit) {
+                        Pull.raiseError(MalformedMessageBodyFailure("Parts limit exceeded"))
+                      } else {
+                        Pull.done
+                      }
+                    } else {
+                      tailrecPartsFileStream[F](
+                        b,
+                        hdrStream,
+                        remaining,
+                        expectedBytes,
+                        headerLimit,
+                        maxBeforeWrite,
+                        partsCounter + 1,
+                        partsLimit,
+                        failOnLimit)
+                        .handleErrorWith(e => cleanupFileOption(fileRef) >> Pull.raiseError(e))
+                    }
+                }
+            }
+        }
+      }
+
+  private[this] def makePart[F[_]](hdrs: Headers, body: Stream[F, Byte], path: Option[Path])(
+      implicit F: Sync[F]): Part[F] = path match {
+    case Some(p) => Part(hdrs, body.onFinalize(F.delay(Files.delete(p))))
+    case None => Part(hdrs, body)
+  }
+
+  /** Split the stream on `values`, but when
+    */
+  //noinspection ScalaStyle
+  private def splitWithFileStream[F[_]](
       values: Array[Byte],
       stream: Stream[F, Byte],
-      limit: Int): Pull[F, Nothing, (Stream[F, Byte], Stream[F, Byte])] = {
+      maxBeforeWrite: Int)(implicit F: Sync[F]): SplitFileStream[F] = {
+
+    def streamAndWrite(
+        s: Stream[F, Byte],
+        state: Int,
+        lacc: Stream[F, Byte],
+        racc: Stream[F, Byte],
+        limitCTR: Int,
+        fileRef: Path): SplitFileStream[F] =
+      if (state == values.length) {
+        Pull.eval(
+          lacc
+            .through(io.file.writeAll[F](fileRef, List(StandardOpenOption.APPEND)))
+            .compile
+            .drain) >> Pull.pure(
+          (io.file.readAll[F](fileRef, maxBeforeWrite), racc ++ s, Some(fileRef)))
+      } else if (limitCTR >= maxBeforeWrite) {
+        Pull.eval(
+          lacc
+            .through(io.file.writeAll[F](fileRef, List(StandardOpenOption.APPEND)))
+            .compile
+            .drain) >> streamAndWrite(s, state, Stream.empty, racc, 0, fileRef)
+      } else {
+        s.pull.unconsChunk.flatMap {
+          case Some((chnk, str)) =>
+            val (ix, l, r, add) = splitOnChunkLimited[F](values, state, chnk, lacc, racc)
+            streamAndWrite(str, ix, l, r, limitCTR + add, fileRef)
+          case None =>
+            Pull.eval(F.delay(Files.delete(fileRef)).attempt) >> Pull.raiseError(
+              MalformedMessageBodyFailure("Invalid boundary - partial boundary"))
+        }
+      }
 
     def go(
         s: Stream[F, Byte],
         state: Int,
         lacc: Stream[F, Byte],
         racc: Stream[F, Byte],
-        limitCTR: Int): Pull[F, Nothing, (Stream[F, Byte], Stream[F, Byte])] =
-      if (limitCTR >= limit) {
-        Pull.raiseError(
-          MalformedMessageBodyFailure(s"Part header was longer than $limit-byte limit"))
+        limitCTR: Int): SplitFileStream[F] =
+      if (limitCTR >= maxBeforeWrite) {
+        Pull
+          .eval(F.delay(Files.createTempFile("", "")))
+          .flatMap { path =>
+            (for {
+              _ <- Pull.eval(lacc.through(io.file.writeAll[F](path)).compile.drain)
+              split <- streamAndWrite(s, state, Stream.empty, racc, 0, path)
+            } yield split)
+              .handleErrorWith(e => Pull.eval(cleanupFile(path)) >> Pull.raiseError(e))
+          }
       } else if (state == values.length) {
-        Pull.pure((lacc, racc ++ s))
+        Pull.pure((lacc, racc ++ s, None))
       } else {
         s.pull.unconsChunk.flatMap {
           case Some((chnk, str)) =>
             val (ix, l, r, add) = splitOnChunkLimited[F](values, state, chnk, lacc, racc)
             go(str, ix, l, r, limitCTR + add)
           case None =>
-            if (state != 0) {
-              Pull.raiseError(MalformedMessageBodyFailure("Invalid boundary - partial boundary"))
-            } else {
-              Pull.pure((lacc, racc))
-            }
+            Pull.raiseError(MalformedMessageBodyFailure("Invalid boundary - partial boundary"))
         }
       }
 
@@ -505,7 +779,7 @@ object MultipartParser {
           splitOnChunkLimited[F](values, 0, chunk, Stream.empty, Stream.empty)
         go(rest, ix, l, r, add)
       case None =>
-        Pull.pure((Stream.empty, Stream.empty))
+        Pull.raiseError(MalformedMessageBodyFailure("Invalid boundary - partial boundary"))
     }
   }